pyproject.toml
[project]
name = "chatbot-project-wim2"
version = "0.1.0"
description = "chatbot project"
authors = [
    {name = "Daniel Lambrecht",email = "daniel.lambrecht@student.reutlingen-university.de"}
]
package-mode = false
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8,<3.11"
dependencies = [
    "rasa (>=3.6.21,<4.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
